+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonom_2Stack_Rosu_Fata.java
@@ -1,6 +1,8 @@
 package org.firstinspires.ftc.teamcode;
 
 import static org.firstinspires.ftc.teamcode.Var_Red.CV_detectionType;
+import static org.firstinspires.ftc.teamcode.Var_Red.Webcam_h;
+import static org.firstinspires.ftc.teamcode.Var_Red.Webcam_w;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
@@ -15,17 +17,14 @@ import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
 import org.openftc.easyopencv.OpenCvCameraRotation;
-
 @Autonomous
-public class Autonom_Rosu_De_Pixel extends LinearOpMode {
+public class Autonom_2Stack_Rosu_Fata extends LinearOpMode {
     double rectx, recty, hperw,x;
     boolean outOfThread = false;
-    int varrez = 2;
-    long lastTime;
+    int varrez = 2,poz2=100;
+    public ChestiiDeAutonom c = new ChestiiDeAutonom(this);
     public OpenCvCamera webcam;
     public PachetelNouRosu pipelineRosu = new PachetelNouRosu(this);
-    public ChestiiDeAutonom c = new ChestiiDeAutonom(this);
-
     @Override
     public void runOpMode() throws InterruptedException {
         c.init(hardwareMap);
@@ -37,7 +36,7 @@ public class Autonom_Rosu_De_Pixel extends LinearOpMode {
         webcam.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener() {
             @Override
             public void onOpened() {
-                webcam.startStreaming(Var_Red.Webcam_w, Var_Red.Webcam_h, OpenCvCameraRotation.UPRIGHT);
+                webcam.startStreaming(Webcam_w, Webcam_h, OpenCvCameraRotation.UPRIGHT);
             }
             @Override
             public void onError(int errorCode) {
@@ -79,153 +78,163 @@ public class Autonom_Rosu_De_Pixel extends LinearOpMode {
         Pose2d startPose = new Pose2d(14.783464, -62.73622, Math.toRadians(90));
         drive.setPoseEstimate(startPose);
         TrajectorySequence ts = drive.trajectorySequenceBuilder(startPose)
-                .addDisplacementMarker(() -> new Thread(() ->{
-                    c.melctarget(0.85, 1300, 3000);
+                .lineTo(new Vector2d(0, 0))
+                .build();
+        if(varrez == 3) {
+            ts = drive.trajectorySequenceBuilder(startPose)
+                    .lineToLinearHeading(new Pose2d(new Vector2d(14, -38), Math.toRadians(45)))
+                    .addTemporalMarker(1, 0, () -> new Thread(() -> {
+                        c.kdf(1500);
+                        c.melctarget(0.8, 1300, 3000);
+                        c.target(-800, 1300, c.getSlider(), 3000, 10);
+                        c.kdf(300);
+                        c.target(-250, 1300, c.getSlider(), 3000, 10);
+                        c.kdf(500);
+                    }).start())
+                    .build();
+        }
+        if (varrez == 2) {
+            ts = drive.trajectorySequenceBuilder(startPose)
+                    .lineTo(new Vector2d(14, -36))
+                    .addTemporalMarker(1, 0, () -> new Thread(() -> {
+                        c.melctarget(0.8, 1300, 3000);
+                        c.target(-800, 1300, c.getSlider(), 3000, 10);
+                        c.kdf(300);
+                        c.target(-300, 1300, c.getSlider(), 3000, 10);
+                        c.kdf(500);
+                    }).start())
+                    .build();
+        }
+        else if (varrez == 1) {
+            ts = drive.trajectorySequenceBuilder(startPose)
+                    .lineToLinearHeading(new Pose2d(new Vector2d(15, -48), Math.toRadians(130)))
+                    .lineToLinearHeading(new Pose2d(new Vector2d(12, -38), Math.toRadians(160)))
+                    .addTemporalMarker(1, 0, () -> new Thread(() -> {
+                        c.melctarget(0.8, 1300, 3000);
+                        c.target(-800, 1300, c.getSlider(), 3000, 10);
+                        c.kdf(300);
+                        c.target(-300, 1300, c.getSlider(), 3000, 10);
+                        c.kdf(500);
+                    }).start())
+                    .build();
+        }
+        drive.followTrajectorySequence(ts);
+        ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                .lineToLinearHeading(new Pose2d(new Vector2d(52, -28), Math.toRadians(180)))
+                .build();
+        if (varrez == 3) {
+            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                    .lineToLinearHeading(new Pose2d(new Vector2d(16, -48), Math.toRadians(90)))
+                    .lineToLinearHeading(new Pose2d(new Vector2d(52, -42), Math.toRadians(180)))
+                    .build();
+        }
+        else if (varrez == 2) {
+            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                    .lineToLinearHeading(new Pose2d(new Vector2d(52, -35), Math.toRadians(180)))
+                    .build();
+        }
+        if (!isStopRequested()) {
+            drive.followTrajectorySequence(ts);
+        }
+        long lastTime = System.currentTimeMillis();
+        while(lastTime + 600 > System.currentTimeMillis() && !isStopRequested()) {
+            c.deschidere();
+        }
+        c.inchidere();
+        ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                .addDisplacementMarker(() -> new Thread(() -> {
+                    c.melctargetRealAngle(440,1000,3000);
+                }).start())
+                .waitSeconds(1)
+                .splineTo(new Vector2d(10,-58),Math.toRadians(180))
+                .addDisplacementMarker(() -> new Thread(() -> {
+                    c.setExtensorPower(1,2000);
+                }).start())
+                .lineToLinearHeading(new Pose2d(new Vector2d(-30, -58), Math.toRadians(180)))
+                .addDisplacementMarker(() -> new Thread(() -> {
+                    c.operation_pixel();
+                }).start())
+                .splineTo(new Vector2d(-56,-44),Math.toRadians(140))
+                .build();
+        if (!isStopRequested()) {
+            drive.followTrajectorySequence(ts);
+            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                    .lineToLinearHeading(new Pose2d(new Vector2d(-54,-37),Math.toRadians(140)))
+                    .build();
+            drive.followTrajectorySequence(ts);
+        }
+        ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                .setReversed(true)
+                .splineTo(new Vector2d(-30,-58),Math.toRadians(0))
+                .lineToSplineHeading(new Pose2d(new Vector2d(10,-58),Math.toRadians(180)))
+                .addTemporalMarker(1, 0, () -> new Thread(() -> {
+                    c.kdf(1000);
+                    c.melctarget(0.85, 700, 3000);
                     c.target(-800, 1300, c.getSlider(), 3000, 10);
                     c.kdf(300);
                     c.target(-300, 1300, c.getSlider(), 3000, 10);
                     c.kdf(500);
                 }).start())
-                .lineToLinearHeading(new Pose2d(new Vector2d(14, -38),Math.toRadians(45)))
                 .build();
-        if(varrez == 2){
-            ts = drive.trajectorySequenceBuilder(startPose)
-                    .lineToLinearHeading(new Pose2d(new Vector2d(14, -36),Math.toRadians(90)))
-                    .addTemporalMarker(0.5,0,() -> new Thread(() -> {
-                        c.melctarget(0.85, 1300, 3000);
-                        c.target(-800, 1300, c.getSlider(), 3000, 10);
-                        c.kdf(300);
-                        c.target(-300, 1300, c.getSlider(), 3000, 10);
-                        c.kdf(500);
-                    }).start())
-                    .build();
+        if (!isStopRequested()) {
+            drive.followTrajectorySequence(ts);
         }
-        else if(varrez == 1){
-            ts = drive.trajectorySequenceBuilder(startPose)
-                    .lineToLinearHeading(new Pose2d(new Vector2d(15,-48),Math.toRadians(130)))
-                    .lineToLinearHeading(new Pose2d(new Vector2d(12,-38),Math.toRadians(160)))
+        if(varrez == 2 || varrez == 1){
+            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                    .setReversed(true)
+                    .splineTo(new Vector2d(52.5,-41),Math.toRadians(0))
                     .addTemporalMarker(1,0,() -> new Thread(() -> {
-                        c.melctarget(0.85, 1300, 3000);
-                        c.target(-800, 1300, c.getSlider(), 3000, 10);
-                        c.kdf(300);
-                        c.target(-300, 1300, c.getSlider(), 3000, 10);
-                        c.kdf(500);
+                        c.setExtensorPower(-1,2000);
                     }).start())
                     .build();
         }
-        drive.followTrajectorySequence(ts);
-        ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                .lineToLinearHeading(new Pose2d(new Vector2d(53, -29),Math.toRadians(180)))
-                .addTemporalMarker(1, 0, () -> new Thread(() -> {
-                    c.setExtensorPower(1,2000);
-                }).start())
-                .build();
-        if(varrez == 3){
+        else{
             ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(new Vector2d(16,-48),Math.toRadians(90)))
-                    .lineToLinearHeading(new Pose2d(new Vector2d(53,-41),Math.toRadians(180)))
-                    .addTemporalMarker(1, 0, () -> new Thread(() -> {
-                        c.setExtensorPower(1,2000);
+                    .setReversed(true)
+                    .splineTo(new Vector2d(52.5,-37),Math.toRadians(0))
+                    .addTemporalMarker(1,0,() -> new Thread(() -> {
+                        c.setExtensorPower(-1,2000);
                     }).start())
                     .build();
         }
-        else if(varrez == 2){
-            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(new Vector2d(53,-34),Math.toRadians(180)))
-                    .addTemporalMarker(1, 0, () -> new Thread(() -> {
-                        c.setExtensorPower(1,2000);
-                    }).start())
-                    .build();
-        }
-        drive.followTrajectorySequence(ts);
-        lastTime = System.currentTimeMillis();
-        while(lastTime + 600 > System.currentTimeMillis() && !isStopRequested()) {
-            c.deschidereJumate(); 
-        }
-        c.inchidere();
-        ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                .addDisplacementMarker(() -> new Thread(() -> {
-                    c.melctarget(2.05,1300,3000);
-                }).start())
-                .waitSeconds(0.5)
-                .lineToLinearHeading(new Pose2d(new Vector2d(37,-10.5),Math.toRadians(180)))
-                .addDisplacementMarker(() -> new Thread(() -> {
-                    c.operation_pixel();
-                }).start())
-                .lineToLinearHeading(new Pose2d(new Vector2d(-52,-10.5),Math.toRadians(180)))
-                .build();
-        drive.followTrajectorySequence(ts);
-        c.kdf(700);
-        ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                .lineTo(new Vector2d(30,-11))
-                .addTemporalMarker(1, 0, () -> new Thread(() -> {
-                    c.melctarget(0.9, 1300, 3000);
-                    c.target(-700, 1300, c.getSlider(), 3000, 10);
-                    c.kdf(300);
-                    c.setExtensorPower(-1,2000);
-                }).start())
-                .build();
-        drive.followTrajectorySequence(ts);
-        ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                .lineTo(new Vector2d(52.5,-34))
-                .build();
-        if(varrez == 2){
-            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                    .lineTo(new Vector2d(52.5,-30))
-                    .build();
-        }
-        else if(varrez == 1){
-            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
-                    .lineTo(new Vector2d(52.5,-41))
-                    .build();
-        }
-        drive.followTrajectorySequence(ts);
-        c.kdf(1000);
-        lastTime = System.currentTimeMillis();
-        while(lastTime + 500 > System.currentTimeMillis() && !isStopRequested()) {
+        if (!isStopRequested()) {
+            drive.followTrajectorySequence(ts);
+            c.kdf(2300);
+            c.deschidereJumate();
+            c.kdf(200);
             c.deschidere();
+            c.kdf(500);
+            c.inchidere();
+            c.melctargetRealAngle(440, 3000, 3000);
+            ts = drive.trajectorySequenceBuilder(drive.getPoseEstimate())
+                    .setReversed(true)
+                    .lineTo(new Vector2d(47,-60))
+                    .build();
+            drive.followTrajectorySequence(ts);
         }
-        c.inchidere();
-        c.melctarget(2.3,3000,3000);
     }
-
     private final Thread Pus_pe_tabla = new Thread(new Runnable() {
         @Override
         public void run() {
-            outOfThread = false;
-            c.melctarget(0.9,1300,3000);
-            c.target(-800,1300,c.getSlider(),3000,10);
-            c.kdf(200);
-            c.target(-200,1300,c.getSlider(),3000,10);
-            c.kdf(500);
-
-            outOfThread = true;
+            if(!isStopRequested()) {
+                outOfThread = false;
+                c.melctarget(0.8, 1300, 3000);
+                c.target(-800, 1300, c.getSlider(), 3000, 10);
+                c.kdf(300);
+                c.target(-300, 1300, c.getSlider(), 3000, 10);
+                c.kdf(500);
+                outOfThread = true;
+            }
         }
     });
     private final Thread Brat_jos = new Thread(new Runnable() {
         @Override
         public void run() {
-            c.inchidere();
-            c.kdf(500);
-            c.melctarget(2.05,1300,3000);
-            c.setMacetaPower(-1);
+            if(!isStopRequested()) {
+                c.inchidere();
+                c.kdf(500);
+                c.melctarget(2.3, 1300, 3000);
+            }
         }
     });
-    public void untilLine (SampleMecanumDrive mecanumDrive)
-    {
-        while(true)
-        {
-            if(gamepad1.x)
-                break;
-        }
-        Pose2d startPose = new Pose2d(14.783464, -62.73622, Math.toRadians(90));
-        mecanumDrive.setPoseEstimate(startPose);
-        mecanumDrive.setMotorPowers(-0.4,0.4,-0.4,0.4);
-        while(!c.isColorWhite())
-        {
-            mecanumDrive.update();
-        }
-        mecanumDrive.update();
-        mecanumDrive.setMotorPowers(0,0,0,0);
-    }
 }
