+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonom_Rosu_De_Balta.java
@@ -134,8 +134,8 @@ public class Autonom_Rosu_De_Balta extends LinearOpMode {
         drive.followTrajectorySequence(ts);
         c.kdf(2000);
         long lastTime = System.currentTimeMillis();
-        while(lastTime + 500 < System.currentTimeMillis() && !isStopRequested()) {
-            c.deschidereJumate();
+        while(lastTime + 500 > System.currentTimeMillis() && !isStopRequested()) {
+            c.deschidere();
         }
         Brat_stop.start();
         if(varrez == 1){
