+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TestTSEdetection.java
@@ -6,13 +6,14 @@ import static org.firstinspires.ftc.teamcode.Var_Blue.Webcam_w;
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
 import org.openftc.easyopencv.OpenCvCameraRotation;
-
+@Disabled
 @Autonomous
 public class TestTSEdetection extends LinearOpMode {
     double rectx, recty, hperw;
